Add BLIP and text-to-image serving (#181)

+++ lib/bumblebee/multimodal/clip.ex
@@ -39,12 +39,10 @@ defmodule Bumblebee.Multimodal.Clip do
       padding tokens, which are added when processing a batch of sequences
       with different length.
 
-
     * `"position_ids"` - `{batch_size, sequence_length}`
 
       Indices of positions of each input sequence tokens in the position
       embeddings.
-
 
     * `"pixel_values"` - `{batch_size, image_size, image_size, num_channels}`
 
@@ -83,7 +81,7 @@ defmodule Bumblebee.Multimodal.Clip do
 
     %{
       "input_ids" => Nx.template({1, 1}, :s64),
-      "pixel_values" => Nx.template(vision_shape, :s64)
+      "pixel_values" => Nx.template(vision_shape, :f32)
     }
   end
